+++ src/main/gen/dev/xeonkryptos/xeonrobotframeworkplugin/psi/element/RobotTemplateArgument.java
@@ -7,7 +7,4 @@ import com.intellij.psi.PsiElement;
 
 public interface RobotTemplateArgument extends RobotArgument, RobotElement {
 
-  @Nullable
-  RobotVariable getVariable();
-
 }